lab3_rag/flask_app/requirements.txt
flask==3.1.0
gunicorn==23.0.0
minio==7.2.12
chromadb-client==0.6.3
sentence-transformers==3.3.1
pdfplumber==0.11.4
python-docx==1.1.2
requests==2.32.3
numpy>=1.26.0
rank-bm25==0.2.2
